polymod0: provide polymodulus as an argument rather than using a global
For quickchecks we will want to sometimes mod out by different
polynomials rather than the main codex32 one. (Also true for "long
codex32" in case we ever try to implement that in postscript..).

Also slightly cleans up the formatting of the polymod0 function, though
I don't try to clean up the logic (which seemed difficult and
unnecessary to untangle).